Arduino 0.5 Hz Sine waveform pump controller
This is the Arduino PID algorithm to control pump to produce Sine wave flow rate with frequency of 0.5 Hz. This code don't uses gain scheduling to produce the desired flow rate because of simple flow rate.
After downloading, the file should be in a folder with same name as filename.