Arduino 1 Hz Sine wave pump controller
This is the Arduino PID algorithm to control a pump to produce Sine wave flow rate with frequency of 1 Hz. This code uses gain scheduling to produce the desired flow rate accurately because 1 Hz signal varies fast and needs different gains for different part of the cycle.
After downloading, the file should be in a folder with same name as filename.