Arduino 0.5 Hz multiple Sine waves pump controller
This is the Arduino PID algorithm to control a pump to produce summation of three sine waves flow rate with total frequency of 0.5 Hz. This code uses gain scheduling to produce the desired flow rate accurately because this 0.5 Hz signal varies fast and needs different gains for different part of the cycle.
After downloading, the file should be in a folder with same name as filename.