Arduino pump controller program to produce common carotid artery flow rate
This Arduino PID algorithm controls a pump to produce physiological  common carotid artery flow rate with frequency of 0.25 Hz. This code uses gain scheduling to produce the desired flow rate accurately because this signal varies fast with large magnitude changes and needs different gains for different part of the cycle.
After downloading, the file should be in a folder with same name as filename.